# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AURA Dealership Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /finance-applications:
    get:
      operationId: listFinanceApplications
      tags: [finance]
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: List of finance applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinanceApplication"
    post:
      operationId: createFinanceApplication
      tags: [finance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceApplication"
  /finance-applications/{id}:
    get:
      operationId: getFinanceApplication
      tags: [finance]
      summary: Application detail with documents and LOS submission history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Application detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceApplicationDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateFinanceApplication
      tags: [finance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceApplicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceApplication"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /finance-applications/{id}/submit:
    post:
      operationId: submitFinanceApplication
      tags: [finance]
      summary: Submit the application to the lender via the LOS connector
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Submitted application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceApplication"
        "400":
          description: Not submittable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: LOS connector failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /finance-applications/{id}/sync:
    post:
      operationId: syncFinanceApplication
      tags: [finance]
      summary: Poll the LOS for the latest decision status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Application after sync
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceApplication"
        "400":
          description: Not submitted yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: LOS connector failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /finance-applications/{id}/documents:
    get:
      operationId: listFinanceDocuments
      tags: [finance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinanceDocument"
    post:
      operationId: uploadFinanceDocument
      tags: [finance]
      summary: Upload a supporting document (multipart form; consumed via raw fetch, not a generated hook)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Uploaded document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceDocument"
        "400":
          description: Invalid upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /finance-applications/{id}/documents/{docId}:
    delete:
      operationId: deleteFinanceDocument
      tags: [finance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: docId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /finance-applications/{id}/documents/{docId}/download:
    get:
      operationId: downloadFinanceDocument
      tags: [finance]
      summary: Download / view a finance document
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: docId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: File contents
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /banks:
    get:
      operationId: listBanks
      tags: [finance]
      responses:
        "200":
          description: Bank registry
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bank"
    post:
      operationId: createBank
      tags: [finance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bank"
  /banks/{id}:
    patch:
      operationId: updateBank
      tags: [finance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bank"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /invoices:
    get:
      operationId: listInvoices
      tags: [finance]
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [finance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    patch:
      operationId: updateInvoice
      tags: [finance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    FinanceApplication:
      type: object
      required: [id, customerName, amount, downPayment, termMonths, apr, status, statusHistory, protectionProducts, createdAt]
      properties:
        id: { type: integer }
        dealId: { type: ["integer", "null"] }
        leadId: { type: ["integer", "null"] }
        customerId: { type: ["integer", "null"] }
        customerName: { type: string }
        amount: { type: number }
        downPayment: { type: number }
        termMonths: { type: integer }
        apr: { type: number }
        lender: { type: ["string", "null"] }
        bankId: { type: ["integer", "null"] }
        employerName: { type: ["string", "null"] }
        jobTitle: { type: ["string", "null"] }
        employmentType: { type: ["string", "null"], enum: [employed, self_employed, contract, retired, other, null] }
        employmentYears: { type: ["number", "null"] }
        monthlyIncome: { type: ["number", "null"] }
        otherIncome: { type: ["number", "null"] }
        status: { type: string, enum: [pending, submitted, under_review, approved, declined, disbursed] }
        statusHistory:
          type: array
          items: { $ref: "#/components/schemas/FinanceStatusEvent" }
        losConnector: { type: ["string", "null"] }
        losReference: { type: ["string", "null"] }
        submittedAt: { type: ["string", "null"], format: date-time }
        decisionAt: { type: ["string", "null"], format: date-time }
        disbursedAt: { type: ["string", "null"], format: date-time }
        protectionProducts:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
...
    FinanceApplicationDetail:
      type: object
      required: [application, documents, submissions]
      properties:
        application: { $ref: "#/components/schemas/FinanceApplication" }
        documents:
          type: array
          items: { $ref: "#/components/schemas/FinanceDocument" }
        submissions:
          type: array
          items: { $ref: "#/components/schemas/LosSubmission" }
...